divided the js files into proper modules
