+++ registerRoute.js
@@ -1,0 +1,23 @@
+const express=require("express");
+const router=express.Router();
+
+const {createUser} = require("./dbServer.js")
+
+// Define route for user registration
+
+router.post("/reg", async function (req, res) {
+	console.log("Received POST request to /register");
+	try {
+
+		console.log(req.body);
+		const { username, password } = req.body;
+		console.log("Credentials received:", username, password);
+		await createUser(username, password);
+		res.send("User created successfully");
+	} catch (error) {
+		console.error("Error creating user:", error);
+		res.status(500).send("An error occurred");
+	}
+});
+
+module.exports=router;
