+++ registration.html
@@ -17,7 +17,7 @@
 	</head>
 	<body class="d-flex align-items-center py-4 bg-body-tertiary">
 		<main class="form-signin w-100 m-auto">
-			<form action="/register" method="post">
+			<form action="/reg" method="post">
 				
 				<h1 class="h3 mb-3 fw-normal">Please sign in</h1>
